ChatApp/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
websockets==14.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
qdrant-client>=1.12.0
